+++ SomosMasApp/SomosMasApp/Assets.xcassets/Image_1.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "fotografia_curriculum_linkedin_corporativa_madrid_011.jpeg",
+      "filename" : "Image_1.jpeg",
       "idiom" : "universal",
       "scale" : "1x"
     },
