+++ SomosMasApp/SomosMasApp/Assets.xcassets/Image_2.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Fotografía-para-CV-en-Madrid.jpeg",
+      "filename" : "Image_2.jpeg",
       "idiom" : "universal",
       "scale" : "1x"
     },
