+++ SomosMasApp/SomosMasApp/Assets.xcassets/Image_3.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "marketing.jpeg",
+      "filename" : "Image_3.jpeg",
       "idiom" : "universal",
       "scale" : "1x"
     },
