+++ SomosMasApp/SomosMasApp/Assets.xcassets/Image_4.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "tesorera.jpeg",
+      "filename" : "Image_4.jpeg",
       "idiom" : "universal",
       "scale" : "1x"
     },
