+++ SomosMasApp/SomosMasApp/Assets.xcassets/Image_5.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "e7b59c_5ec00be7b5ea4a59b6b41e48dec9173f_mv2.jpeg",
+      "filename" : "Image_5.jpeg",
       "idiom" : "universal",
       "scale" : "1x"
     },
